requirements.txt
aws-cdk-lib==2.162.1
constructs==10.4.2
aws-solutions-constructs.aws-s3-stepfunctions==2.72.0


